Fix frame data validation and update animation data format
Animation data should include the raw frame data instead of just a
list of frame references. This way it will be possible to create an
animation with a set of frames with a single call to the contract.

Frames that already exist are skipped and frames that don't already
exist will be created by calling the `creatFrame` function.