Copy memory to create rect elements
This is a new implementation of the function to generate the NFT
metadata that fixes the previous issue about the contract size.

The new implementation copies the same template rect element for
each pixel in the resulting SVG.

The performance is still acceptable and the contract size is back
to normal.