Initial implementation of SVG animations
This is an initial implementation of the method to encode animations
as SVG data.

Each frame is encoded using the `_encodeBitmap` function and stored
in the `defs` section of the SVG.

The animation is a `use` element that wraps an `animate` element that
animates the `href` attribute with all the values defined in the
`defs` section.

Note: the contract needs to be optimized because right now it's too
large to be deployed. The `encodeAnimation` method needs also to be
further optimized because it fails to complete with a time out error
when it's used to encode a "large" animations (16 frames).

References:

* https://developer.mozilla.org/en-US/docs/Web/SVG/Element/defs
* https://developer.mozilla.org/en-US/docs/Web/SVG/Element/use
* https://developer.mozilla.org/en-US/docs/Web/SVG/Element/animate
* https://github.com/w3c/svgwg/issues/130
* https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/xlink:href
* https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/repeatCount