Replace width and height rect attributes with CSS
Instead of defining a constant `width` and `height` attribute for each
`rect` element, we can embed a `style` element in the SVG that defines
a global style sheet for the SVG with a constant `width` and `height`
for all `rect` elements.

Note: I'm using `px` as the unit even if the that might be different
from the default unit used by SVG in case a "length" number doesn't
specify any unit.

References:

* https://developer.mozilla.org/en-US/docs/Web/SVG/Element/style

This optimization has been suggested by @bugduino