Add contract-sizer
StunningPotato is too large to be deployed on mainnet. Contract-sizer
is a hardhat plugin that reports basic info about contract size.

The issue is that the report doesn't include any additional information
about contract size other than the total contract size, so in order to
understand what is contributing the most on the size, I need to
manually add and remove functions :(

For instance the current contract implementation size is: 25.858
If I remove the call to `encodeAnimation`: 22.379
If I remove the call to `encodeFrame`: 25.500
If I remove the call to both: 18.802

I guess the reason why removing just `encodeFrame` doesn't reduce
the size by much is because `encodeAnimation` uses `encodeFrame`.

I think there are at least a couple of possible optimizations:

1. Shorten error string size, replace with error codes
2. Store SVG raw code and implement URL encoding function

See https://github.com/ItsNickBarry/hardhat-contract-sizer