Change Solidity optimizer settings
It's a bit counterintuitive to me, but increasing the number of runs
reduced both code size and gas cost.

The best runs value I've found so far is 1000. It reduces the contract
code size to 15.712, that's about 10kb less than with the default
value of 200. The interesting part is that it also reduces the gas
cost of the contract functions.

Resources:

* https://hardhat.org/guides/compile-contracts.html#configuring-the-compiler
* https://docs.soliditylang.org/en/v0.8.12/using-the-compiler.html#optimizer-options